Consolidate haversine functions into a single definition in geo.py
